Extract SwitcherTableRow
